invoke `DimensionManager.setWorld` with null world to force removing invalid world mapping from Magma cache
This should resolve #116, #110, and #102.

Magma injects into `DimensionManager.setWorld` to keep track of their own Bukkit world list. Our system of directly operating on Forge's maps was not sufficient for updating this list, nor was the DimensionManager.setWorld call due to Magma not handle a world being replaced, and instead just adding the compatibility world to the list.